requirements.txt
Flask
pytesseract
Pillow
tqdm
numpy
scikit-learn
sentence-transformers
